requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0
lxml==5.1.0
gunicorn==21.2.0

